+++ data/dsm/dsm.sql
@@ -1,0 +1,6 @@
+CREATE TABLE IF NOT EXISTS dsm (
+   x varchar(32) not null,
+   y varchar(32) not null,
+   z varchar(32) not null,
+   sig_cd character varying(14)
+)
