+++ data/dsm/testGridPoints.txt
@@ -1,0 +1,6 @@
+128.0001388888 37.99986111111 226346.49284393355 336153.93131701986 354211 574118
+226353.12643123957 335749.2824913528 226565.4012250321 336153.93131701986 354211 574028
+226558.7676377261 335656.41226906853 226810.84395535474 336140.6641424078 354211 573938
+226061.2485897748 336412.6412219545 226359.76001854558 336797.38928570354 354211 573848
+226379.66078046363 336352.93893620034 226618.46992348024 336757.5877618674 354211 573758
+226618.46992348024 336359.57252350636 226817.47754266075 336691.2518888072 354211 573668
