+++ src/main/java/module/DsmInitializer.java
@@ -1,10 +1,15 @@
-import config.ApplicationProperties;
+package module;
 
-public class main {
+import config.ApplicationProperties;
+import service.DsmService;
+
+public class DsmInitializer {
 
     private static final ApplicationProperties applicationProperties = new ApplicationProperties();
 
     public static void main(String[] args) {
-
+        DsmService dsmService = new DsmService();
+        dsmService.storeDsm();
     }
+
 }
