+++ src/main/java/module/RoadInitializer.java
@@ -1,10 +1,15 @@
-import config.ApplicationProperties;
+package module;
 
-public class main {
+import config.ApplicationProperties;
+import service.RoadService;
+
+public class RoadInitializer {
 
     private static final ApplicationProperties applicationProperties = new ApplicationProperties();
 
     public static void main(String[] args) {
-
+        RoadService roadService = new RoadService();
+        roadService.storeRoad();
     }
+
 }
