--- src/main/java/storeDsm/files/dsm.sql
@@ -1,6 +1,0 @@
-CREATE TABLE IF NOT EXISTS dsm (
-   x varchar(32) not null,
-   y varchar(32) not null,
-   z varchar(32) not null,
-   sig_cd numeric default 0
-)
