+++ src/main/resources/application.properties
@@ -3,13 +3,9 @@ jdbc.user = postgres
 jdbc.password = 1
 jdbc.driver = org.postgresql.Driver
 
-dsm.path = src/main/java/storeDsm/files
-dsm.tableName = dsm
-<<<<<<< HEAD
-dsm.createFilePath = src/main/java/storeDsm/files/dsm.sql
+dsm.path = ./data/dsm/
+dsm.testPath = ./data/dsm/testGridPoints.txt
+dsm.createFilePath = ./data/dsm/dsm.sql
 
-road.createFilePath = src/main/java/road/storeRoad/files/road.sql
-road.shpPath = src/main/java/road/storeRoad/files/shp/
-=======
-dsm.createFilePath = src/main/java/storeDsm/files/dsm.sql
->>>>>>> ec2c922ad3fabece5d0cbe471d6ed66e98ac6338
+road.createFilePath = ./data/road/road.sql
+road.shpPath = ./data/road/shp
